Task#5: refactoring
