+++ client/src/features/MovieDetails/container.connected.js
@@ -1,0 +1,14 @@
+import { connect } from 'react-redux';
+
+import MovieDetailsContainer from './container';
+
+import { getMovie, getMoviesByGenre } from './actions';
+import { movieSelector, moviesByGenreSelector } from 'core/store/selectors';
+
+export default connect(
+  state => ({
+    movie: movieSelector(state),
+    moviesByGenre: moviesByGenreSelector(state),
+  }),
+  { getMovie, getMoviesByGenre },
+)(MovieDetailsContainer);
