+++ client/src/features/SearchForm/container.connected.js
@@ -1,0 +1,9 @@
+import { connect } from 'react-redux';
+
+import SearchFormContainer from './container';
+
+import { searchBySelector } from 'core/store/selectors';
+
+export default connect(state => ({
+  searchBy: searchBySelector(state),
+}))(SearchFormContainer);
