+++ client/src/features/SearchMovie/container.js
@@ -1,10 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
 
-import { allMoviesSelector, sortValueSelector } from 'core/store/selectors';
-
-import { getMovies, searchMovies, sortMovies, setSortBy, setSearchBy } from './actions';
 import SearchMovieComponent from './component';
 
 class SearchMovieContainer extends Component {
@@ -52,10 +48,4 @@ class SearchMovieContainer extends Component {
   }
 }
 
-export default connect(
-  state => ({
-    movies: allMoviesSelector(state),
-    sortBy: sortValueSelector(state),
-  }),
-  { getMovies, searchMovies, sortMovies, setSortBy, setSearchBy },
-)(SearchMovieContainer);
+export default SearchMovieContainer;
