+++ client/src/features/SearchMovie/index.js
@@ -1,2 +1,2 @@
-export { default } from './container';
+export { default } from './container.connected';
 export { default as reducer } from './reducer';
